implemented AbstractionMap to replace JsonSerializationAbstractionMap
