+++ Manatee.Json.Tests/Path/TestSuite/JsonPathTestSuite.cs
@@ -7,49 +7,49 @@ using NUnit.Framework;
 
 namespace Manatee.Json.Tests.Path.TestSuite
 {
-    [TestFixture]
-    public class JsonPathTestSuite
-    {
-        private const string TestFolder = @"..\..\..\Json-Path-Test-Suite\tests\";
-        private static readonly JsonSerializer _serializer;
+	[TestFixture]
+	public class JsonPathTestSuite
+	{
+		private const string TestFolder = @"..\..\..\Json-Path-Test-Suite\tests\";
+		private static readonly JsonSerializer _serializer;
 
-        public static IEnumerable TestData
-        {
-            get
-            {
-                var testsPath = System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, TestFolder).AdjustForOS();
-                var fileNames = Directory.GetFiles(testsPath);
+		public static IEnumerable TestData
+		{
+			get
+			{
+				var testsPath = System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, TestFolder).AdjustForOS();
+				var fileNames = Directory.GetFiles(testsPath);
 
-                foreach (var fileName in fileNames)
-                {
-                    var contents = File.ReadAllText(fileName);
-                    var json = JsonValue.Parse(contents);
+				foreach (var fileName in fileNames)
+				{
+					var contents = File.ReadAllText(fileName);
+					var json = JsonValue.Parse(contents);
 
-                    var testSet = _serializer.Deserialize<PathTestSet>(json);
+					var testSet = _serializer.Deserialize<PathTestSet>(json);
 
-                    foreach (var test in testSet.Tests)
-                    {
-                        yield return new TestCaseData(testSet.Data, test)
-                            {
-                                TestName = $"{testSet.Title} / {test.Path}"
-                            };
-                    }
-                }
-            }
-        }
+					foreach (var test in testSet.Tests)
+					{
+						yield return new TestCaseData(testSet.Data, test)
+							{
+								TestName = $"{testSet.Title} / {test.Path}"
+							};
+					}
+				}
+			}
+		}
 
-        static JsonPathTestSuite()
-        {
-            _serializer = new JsonSerializer();
-            JsonSerializationAbstractionMap.MapGeneric(typeof(IEnumerable<>), typeof(List<>));
-        }
+		static JsonPathTestSuite()
+		{
+			_serializer = new JsonSerializer();
+			_serializer.AbstractionMap.MapGeneric(typeof(IEnumerable<>), typeof(List<>));
+		}
 
-        [TestCaseSource(nameof(TestData))]
-        public void Run(JsonValue data, PathTest test)
-        {
-            var results = test.Path.Evaluate(data);
-            
-            Assert.AreEqual(test.Result, results);
-        }
-    }
+		[TestCaseSource(nameof(TestData))]
+		public void Run(JsonValue data, PathTest test)
+		{
+			var results = test.Path.Evaluate(data);
+			
+			Assert.AreEqual(test.Result, results);
+		}
+	}
 }
