+++ Manatee.Json/Serialization/Internal/Serializers/JsonSerializableSerializer.cs
@@ -11,7 +11,7 @@
 		}
 		public T Deserialize<T>(JsonValue json, JsonSerializer serializer)
 		{
-			var value = (IJsonSerializable) JsonSerializationAbstractionMap.CreateInstance<T>(json, serializer.Options.Resolver);
+			var value = (IJsonSerializable)serializer.AbstractionMap.CreateInstance<T>(json, serializer.Options.Resolver);
 			value.FromJson(json, serializer);
 			return (T) value;
 		}
